# v7: optimize evaluation when all rules have the same value