@laliconfigcat 2021-11-03
Hey @rogpeppe,

Thanks for your PR. I really like the idea for improving the performance, however, I have 2 concerns regarding this change:

1. Your change would break the current evaluation logic regarding the percentage rollout.
Let's say you have a ruleset like this:
```
if email isoneof 'a@example.com' then false
Else 20% true / 80% false
default false.
```
Before your change, this ruleset for the a@example.com user would be evaluated as false because of the rule. After your change, this would be evaluated by the percentage rule and maybe it would be true.

This part could be fixed by checking if the feature flag has percentage rules too, and in that case, the rollout rules shouldn't be truncated. This could be improved if you would truncate the percentage rules first and you would check if there are any percentage rules left after the truncation.

2. This change would break our debug logging experience. 
Let's say you have a ruleset like this:
```
if email isoneof 'a@example.com' then false
default false.
```
Before your change the following logs would be created:
Evaluating rule ... => match, returning: false
After your change the following logs would be created:
Returning false.
This would be confusing for those who are trying to figure out why a feature flag was evaluated to a value with our debug logs.
Maybe this could be fixed by e.g. a configuration parameter during the client initialization (defaults to false) that turns on/off the truncating logic so the default behavior could be the same as before.

Also, it would be awesome if you could write some unit tests too for checking the evaluation logic with real scenarios (values are returned correctly) as this is really the most important part of the SDK. I see that you created some benchmark tests, but e.g. the problem mentioned in the 1. point should be tested as well.

What do you think?

Cheers,
Lali