@rogpeppe 2022-02-09
> Hey @rogpeppe,
> 
> Thanks for your PR. I really like the idea for improving the performance, however, I have 2 concerns regarding this change:
> 
> 1. Your change would break the current evaluation logic regarding the percentage rollout.
>    Let's say you have a ruleset like this:
> 
> ```
> if email isoneof 'a@example.com' then false
> Else 20% true / 80% false
> default false.
> ```
> 
> Before your change, this ruleset for the [a@example.com](mailto:a@example.com) user would be evaluated as false because of the rule. After your change, this would be evaluated by the percentage rule and maybe it would be true.

That's an excellent point!

> This part could be fixed by checking if the feature flag has percentage rules too, and in that case, the rollout rules shouldn't be truncated. This could be improved if you would truncate the percentage rules first and you would check if there are any percentage rules after the truncation.

Yes, that's exactly what I just thought. Quite straightforward to do.

> 2. This change would break our debug logging experience.
>    Let's say you have a ruleset like this:
> 
> ```
> if email isoneof 'a@example.com' then false
> default false.
> ```
> 
> Before your change the following logs would be created: Evaluating rule ... => match, returning: false After your change the following logs would be created: Returning false. This would be confusing for those who are trying to figure out why a feature flag was evaluated to a value with our debug logs. Maybe this could be fixed by e.g. a configuration parameter during the client initialization (defaults to false) that turns on/off the truncating logic so the default behavior could be the same as before.

Another possibility: we could do the optimization only if the log level is above Info level (i.e. when we wouldn't see the log messages anyway). From a quick glance at the code (it's been a while :) ), I _think_ that we always know the max log level when we create the evaluation function, so we could just predicate the optimisation on that. So it wouldn't be observable - just speed things up when you've turned off logging (and if you really care about performance, you're going to disable logging anyway - we certainly do).

> 
> Also, it would be awesome if you could write some unit tests too for checking the evaluation logic with real scenarios (values are returned correctly) as this is really the most important part of the SDK. I see that you created some benchmark tests, but e.g. the problem mentioned in the 1. point should be tested as well.

I'd love to write some unit tests. One issue is that a bunch of the unit tests are actually contained in your servers and pulled in automatically as part of the tests. There is a local copy of them (in the `resources` directory) but I obviously don't have the authority to add any test data to them. Ideally we'd add test cases to them for this new logic, but in the meantime, I can add some explicit test cases that aren't part of that suite.

> 
> What do you think?

Thanks for the very swift response! I apologise for my slow turnaround - I must have missed your response in the usual deluge of github emails... What do you think about my suggestions above?

   cheers,
    rog.